// src/main/webapp/src/app/segregator-ts-api/model/resultGitVersion.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GitVersion } from './gitVersion';


export interface ResultGitVersion { 
    error?: number;
    message?: string;
    result?: GitVersion;
}