// src/main/webapp/src/app/segregator-ts-api/model/resultListSegItem.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SegItem } from './segItem';


export interface ResultListSegItem { 
    list?: Array<SegItem>;
    nextPage?: boolean;
    resultSize?: number;
    totalSize?: number;
}